--- addon/templates/frost-list-wrapper.hbs
@@ -1,17 +1,0 @@
-{{frost-list recordComponent
-  items=items
-  selection=selection
-  sorting=sorting
-  expansion=expansion
-}}
-
-
-
-{{!--
-  {{frost-list recordComponent
-    items=items
-    expansion=expansion
-    selection=selection
-    sorting=sorting
-  }}
---}}
