+++ addon/templates/frost-list.hbs
@@ -1,47 +1,6 @@
-{{#if _hasHeader}}
-  <div class="frost-list-header">
-    {{#if sorting}}
-      {{frost-sort
-      hook=(concat hook '-sort')
-      sortableProperties=sorting.sortableProperties
-      sortParams=sorting.activeSorting
-      onChange=sorting.onSort
-      }}
-    {{/if}}
-    {{#if expansion}}
-      {{frost-list-expansion
-      hook=(concat hook '-enpansion')
-      onCollapseAll=expansion.onCollapseAll
-      onExpandAll=expansion.onExpandAll
-      }}
-    {{/if}}
-  </div>
-{{/if}}
-<div class="frost-list-content">
-    <div class='border-top'></div>
-    <div class='container'>
-      {{#vertical-collection
-        content=_records
-        scrollPosition=scrollPosition
-        defaultHeight=defaultHeight
-        alwaysUseDefaultHeight=alwaysUseDefaultHeight
-        key=key
-        idForFirstItem=idForFirstItem
-        firstReached=loadPrevious
-        lastReached=loadNext
-        as |item index|
-      }}
-        {{#if hasBlock}}
-          {{yield item index}}
-        {{else}}
-          {{component
-          recordComponent
-          model=item
-          hook=(concat hook '-item-' index)
-          hookQualifiers=(hash index=index)
-          }}
-        {{/if}}
-      {{/vertical-collection}}
-    </div>
-    <div class='border-bottom'></div>
-</div>
+{{frost-list-core recordComponent
+  items=items
+  selection=selection
+  sorting=sorting
+  expansion=expansion
+}}
