+++ app/components/frost-list-core.js
@@ -1,0 +1,1 @@
+export { default } from 'ember-frost-list/components/frost-list'
