--- app/components/frost-list-wrapper.js
@@ -1,1 +1,0 @@
-export { default } from 'ember-frost-list/components/frost-list-wrapper'
