+++ app/components/frost-list.js
@@ -1,1 +1,1 @@
-export { default } from 'ember-frost-list/components/frost-list'
+export { default } from 'ember-frost-list/components/frost-list-wrapper'
